chore: bump agda + nixpkgs + ghc
