+++ support/nix/dep/Agda/github.json
@@ -3,6 +3,6 @@
   "repo": "agda",
   "branch": "master",
   "private": false,
-  "rev": "0feec2519d580aaf8cc18b791c16790ca738bfd1",
-  "sha256": "0jsmj2rrjnxw9zxklrddh2i96qdqwzrar4r6s725vpjm659ss714"
+  "rev": "5cb475a67135ca4ce42428c6f0294cea58a3ca2b",
+  "sha256": "082clxnzd1jzhc3q282np4hp2bzmr1q0fixpw0zal6ifzimwnp5m"
 }
