+++ support/web/css/components/controls.scss
@@ -1,4 +1,7 @@
-@import "../vars.scss";
+@use "../vars.scss"   as *;
+@use "../colors.scss" as *;
+
+@use "sass:color";
 
 $dropdown-popup-width: 280px;
 $button-size:          32px;
@@ -99,7 +102,7 @@ button.theme-button {
 
   &.theme-button-light, &.theme-button-system {
     background-color: $light-text-bg;
-    border-color:     darken($light-text-bg, 10%);
+    border-color:     color.adjust($light-text-bg, $lightness: -10%);
     color:            $light-text-fg;
   }
