+++ support/web/css/components/popup.scss
@@ -1,7 +1,7 @@
-div.hover-popup, div.hover-popup.sourceCode {
-  &.popup-hidden { display: none; }
-  @include widescreen { display: block; }
+@use "../mixins.scss" as *;
+@use "../vars.scss"   as *;
 
+div.hover-popup, div.hover-popup.sourceCode {
   position: absolute;
   z-index: 100;
 
@@ -13,16 +13,16 @@ div.hover-popup, div.hover-popup.sourceCode {
 
   box-shadow: 2px 2px 6px theme(--shadow);
 
-  >:first-child { margin-block-start: 0; }
-  >:last-child  { margin-block-end: 0; }
-
   padding: 0.3em;
   border: 0;
   margin: 0;
 
-  &:before {
-    display: none;
-  }
+  >:first-child { margin-block-start: 0; }
+  >:last-child  { margin-block-end: 0; }
+
+  &:before            { display: none; }
+  &.popup-hidden      { display: none; }
+  @include widescreen { display: block; }
 }
 
 .text-popup {
