+++ support/web/css/theme.scss
@@ -1,9 +1,10 @@
+@use "colors" as *;
+@use "vars"   as *;
+
 @use "sass:selector";
+@use "sass:color";
 @use "sass:meta";
 @use "sass:map";
-
-@import "vars.scss";
-@import "colors.scss";
 
 $theme: (
   text: (
@@ -65,7 +66,7 @@ $theme: (
     module:      (light: $light-code-pink,   dark: $dark-code-cyan),
     field:       (light: $light-code-purple, dark: $dark-code-pink),
     constructor: (light: $light-code-green,  dark: $dark-code-green),
-    highlight:   (light: #F5DEB3,          dark: #EF444499),
+    highlight:   (light: #F5DEB3,                 dark: #EF444499),
   ),
 
   search-selected: (light: $violet-800, dark: $cyan-900),
@@ -76,7 +77,7 @@ $theme: (
 
   input: (
     bg:     (light: $light-text-bg, dark: $bluegray-700),
-    border: (light: lighten($light-text-fg, 30%), dark: $bluegray-600)
+    border: (light: color.adjust($light-text-fg, $lightness: 30%), dark: $bluegray-600)
   ),
 
   popup-bg:        (light: $light-text-bg, dark: $dark-text-bg),
